Refactor: put test comparison functions in separate file
So they can more easily be reused by other tests (in other files).
