+++ src/MotoROS2_AllControllers.vcxproj
@@ -344,6 +344,7 @@
     <ClCompile Include="ServiceStartTrajMode.c" />
     <ClCompile Include="ServiceSelectMotionTool.c" />
     <ClCompile Include="Tests_CtrlGroup.c" />
+    <ClCompile Include="Tests_TestUtils.c" />
     <ClCompile Include="MotionControl.c" />
     <ClCompile Include="ActionServer_FJT.c" />
     <ClCompile Include="main.c" />
@@ -374,6 +375,7 @@
     <ClInclude Include="ServiceStartTrajMode.h" />
     <ClInclude Include="ServiceSelectMotionTool.h" />
     <ClInclude Include="Tests_CtrlGroup.h" />
+    <ClInclude Include="Tests_TestUtils.h" />
     <ClInclude Include="TimeConversionUtils.h" />
     <ClInclude Include="MotionControl.h" />
     <ClInclude Include="ActionServer_FJT.h" />
